# docs/swagger.yaml
basePath: /
definitions:
  github_com_qwerty7415963_go_be_arbitrage_internal_api.ErrorBody:
    properties:
      code:
        type: string
      details: {}
      message:
        type: string
      request_id:
        type: string
    type: object
...
  github_com_qwerty7415963_go_be_arbitrage_internal_api.Meta:
    properties:
      cursor:
        type: string
      has_more:
        type: boolean
      limit:
        type: integer
      offset:
        type: integer
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
...
  github_com_qwerty7415963_go_be_arbitrage_internal_api.Response:
    properties:
      data: {}
      error:
        $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.ErrorBody'
      meta:
        $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.Meta'
      success:
        type: boolean
    type: object
...
  internal_strategy.CreateStrategyRequest:
    properties:
      config:
        $ref: '#/definitions/internal_strategy.StrategyConfig'
      mode:
        $ref: '#/definitions/internal_strategy.StrategyMode'
      name:
        type: string
      type:
        $ref: '#/definitions/internal_strategy.StrategyType'
    required:
    - mode
    - name
    - type
    type: object
  internal_strategy.StrategyConfig:
    properties:
      custom_params:
        additionalProperties:
          type: string
        type: object
      instruments:
        items:
          type: string
        type: array
      max_leverage:
        type: number
      max_notional_per_trade:
        type: string
      max_position_usd:
        type: string
      min_confidence:
        type: number
      min_net_edge_bps:
        type: integer
      venues:
        items:
          type: string
        type: array
    type: object
  internal_strategy.StrategyDecision:
    properties:
      action:
        type: string
      created_at:
        type: string
      executed_at:
        type: string
      id:
        type: string
      opportunity_id:
        type: string
      status:
        type: string
      strategy_id:
        type: string
    type: object
  internal_strategy.StrategyInstance:
    properties:
      config:
        $ref: '#/definitions/internal_strategy.StrategyConfig'
      created_at:
        type: string
      created_by:
        type: string
      id:
        type: string
      mode:
        $ref: '#/definitions/internal_strategy.StrategyMode'
      name:
        type: string
      status:
        $ref: '#/definitions/internal_strategy.StrategyStatus'
      strategy_type_id:
        type: string
      tenant_id:
        type: string
      updated_at:
        type: string
    type: object
  internal_strategy.StrategyMode:
    enum:
    - PAPER
    - LIVE_MANUAL
    - LIVE_AUTO
    type: string
    x-enum-varnames:
    - StrategyModePaper
    - StrategyModeLiveManual
    - StrategyModeLiveAuto
  internal_strategy.StrategyStatus:
    enum:
    - DRAFT
    - RUNNING
    - PAUSED
    - DISABLED
    - ERROR
    type: string
    x-enum-varnames:
    - StrategyStatusDraft
    - StrategyStatusRunning
    - StrategyStatusPaused
    - StrategyStatusDisabled
    - StrategyStatusError
  internal_strategy.StrategyType:
    enum:
    - PRICE_ARBITRAGE
    - FUNDING_ARBITRAGE
    - BASIS_ARBITRAGE
    type: string
    x-enum-varnames:
    - StrategyTypePriceArb
    - StrategyTypeFundingArb
    - StrategyTypeBasisArb
  internal_strategy.UpdateStrategyRequest:
    properties:
      config:
        $ref: '#/definitions/internal_strategy.StrategyConfig'
      name:
        type: string
      status:
        $ref: '#/definitions/internal_strategy.StrategyStatus'
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@arbitrage.com
    name: API Support
  description: Backend API for arbitrage trading platform
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Arbitrage Platform API
  version: "1.0"
paths:
...
  /api/v1/storage/decisions/strategy/{instance_id}:
    get:
      consumes:
      - application/json
      description: List strategy decisions for a strategy instance
      parameters:
      - description: Strategy Instance ID
        in: path
        name: instance_id
        required: true
        type: string
      - default: 20
        description: Page size
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.Response'
      summary: List strategy decisions
      tags:
      - storage
...
  /api/v1/strategies:
    get:
      description: 'Strategy instances for the caller''s tenant. Requires JWT + admin
        role (group middleware: 401 AUTH-003 without token, 403 AUTH-005 for non-admin).'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/internal_strategy.StrategyInstance'
                  type: array
              type: object
        "401":
          description: AUTH-003 no/invalid token
          schema:
            allOf:
            - $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.Response'
            - properties:
                error:
                  $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.ErrorBody'
              type: object
        "403":
          description: AUTH-005 not admin
          schema:
            allOf:
            - $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.Response'
            - properties:
                error:
                  $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.ErrorBody'
              type: object
        "500":
          description: COMMON-901
          schema:
            allOf:
            - $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.Response'
            - properties:
                error:
                  $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.ErrorBody'
              type: object
      security:
      - BearerAuth: []
      summary: List strategy instances
      tags:
      - strategies
    post:
      consumes:
      - application/json
      description: Creates a strategy instance scoped to the caller's tenant. Requires
        JWT + admin role.
      parameters:
      - description: Strategy definition
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_strategy.CreateStrategyRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_strategy.StrategyInstance'
              type: object
        "400":
          description: COMMON-902 invalid body
          schema:
            allOf:
            - $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.Response'
            - properties:
                error:
                  $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.ErrorBody'
              type: object
        "401":
          description: AUTH-003
          schema:
            allOf:
            - $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.Response'
            - properties:
                error:
                  $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.ErrorBody'
              type: object
        "403":
          description: AUTH-005
          schema:
            allOf:
            - $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.Response'
            - properties:
                error:
                  $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.ErrorBody'
              type: object
        "500":
          description: COMMON-901
          schema:
            allOf:
            - $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.Response'
            - properties:
                error:
                  $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.ErrorBody'
              type: object
      security:
      - BearerAuth: []
      summary: Create strategy instance
      tags:
      - strategies
  /api/v1/strategies/{id}:
    delete:
      description: Deletes a strategy instance by ID. Requires JWT + admin role.
      parameters:
      - description: Strategy instance ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.Response'
            - properties:
                data:
                  additionalProperties:
                    type: string
                  type: object
              type: object
        "400":
          description: COMMON-902
          schema:
            allOf:
            - $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.Response'
            - properties:
                error:
                  $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.ErrorBody'
              type: object
        "401":
          description: AUTH-003
          schema:
            allOf:
            - $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.Response'
            - properties:
                error:
                  $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.ErrorBody'
              type: object
        "403":
          description: AUTH-005
          schema:
            allOf:
            - $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.Response'
            - properties:
                error:
                  $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.ErrorBody'
              type: object
        "500":
          description: COMMON-901
          schema:
            allOf:
            - $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.Response'
            - properties:
                error:
                  $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.ErrorBody'
              type: object
      security:
      - BearerAuth: []
      summary: Delete strategy instance
      tags:
      - strategies
    get:
      description: One strategy instance by ID. Requires JWT + admin role.
      parameters:
      - description: Strategy instance ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_strategy.StrategyInstance'
              type: object
        "400":
          description: COMMON-902 invalid ID
          schema:
            allOf:
            - $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.Response'
            - properties:
                error:
                  $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.ErrorBody'
              type: object
        "401":
          description: AUTH-003
          schema:
            allOf:
            - $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.Response'
            - properties:
                error:
                  $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.ErrorBody'
              type: object
        "403":
          description: AUTH-005
          schema:
            allOf:
            - $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.Response'
            - properties:
                error:
                  $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.ErrorBody'
              type: object
        "404":
          description: COMMON-903
          schema:
            allOf:
            - $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.Response'
            - properties:
                error:
                  $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.ErrorBody'
              type: object
      security:
      - BearerAuth: []
      summary: Get strategy instance
      tags:
      - strategies
    put:
      consumes:
      - application/json
      description: Updates a strategy instance by ID. Requires JWT + admin role.
      parameters:
      - description: Strategy instance ID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_strategy.UpdateStrategyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_strategy.StrategyInstance'
              type: object
        "400":
          description: COMMON-902
          schema:
            allOf:
            - $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.Response'
            - properties:
                error:
                  $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.ErrorBody'
              type: object
        "401":
          description: AUTH-003
          schema:
            allOf:
            - $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.Response'
            - properties:
                error:
                  $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.ErrorBody'
              type: object
        "403":
          description: AUTH-005
          schema:
            allOf:
            - $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.Response'
            - properties:
                error:
                  $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.ErrorBody'
              type: object
        "500":
          description: COMMON-901
          schema:
            allOf:
            - $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.Response'
            - properties:
                error:
                  $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.ErrorBody'
              type: object
      security:
      - BearerAuth: []
      summary: Update strategy instance
      tags:
      - strategies
  /api/v1/strategies/{id}/decisions:
    get:
      description: Recent decisions for one strategy instance. Requires JWT + admin
        role.
      parameters:
      - description: Strategy instance ID
        in: path
        name: id
        required: true
        type: string
      - default: 50
        description: Max rows
        in: query
        minimum: 1
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/internal_strategy.StrategyDecision'
                  type: array
              type: object
        "400":
          description: COMMON-902
          schema:
            allOf:
            - $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.Response'
            - properties:
                error:
                  $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.ErrorBody'
              type: object
        "401":
          description: AUTH-003
          schema:
            allOf:
            - $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.Response'
            - properties:
                error:
                  $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.ErrorBody'
              type: object
        "403":
          description: AUTH-005
          schema:
            allOf:
            - $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.Response'
            - properties:
                error:
                  $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.ErrorBody'
              type: object
        "500":
          description: COMMON-901
          schema:
            allOf:
            - $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.Response'
            - properties:
                error:
                  $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.ErrorBody'
              type: object
      security:
      - BearerAuth: []
      summary: List strategy decisions
      tags:
      - strategies
  /api/v1/strategies/{id}/start:
    post:
      description: Starts an existing strategy instance. Requires JWT + admin role.
      parameters:
      - description: Strategy instance ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.Response'
            - properties:
                data:
                  additionalProperties:
                    type: string
                  type: object
              type: object
        "400":
          description: COMMON-902
          schema:
            allOf:
            - $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.Response'
            - properties:
                error:
                  $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.ErrorBody'
              type: object
        "401":
          description: AUTH-003
          schema:
            allOf:
            - $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.Response'
            - properties:
                error:
                  $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.ErrorBody'
              type: object
        "403":
          description: AUTH-005
          schema:
            allOf:
            - $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.Response'
            - properties:
                error:
                  $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.ErrorBody'
              type: object
        "500":
          description: COMMON-901
          schema:
            allOf:
            - $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.Response'
            - properties:
                error:
                  $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.ErrorBody'
              type: object
      security:
      - BearerAuth: []
      summary: Start strategy instance
      tags:
      - strategies
  /api/v1/strategies/{id}/stop:
    post:
      description: Stops a running strategy instance. Requires JWT + admin role.
      parameters:
      - description: Strategy instance ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.Response'
            - properties:
                data:
                  additionalProperties:
                    type: string
                  type: object
              type: object
        "400":
          description: COMMON-902
          schema:
            allOf:
            - $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.Response'
            - properties:
                error:
                  $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.ErrorBody'
              type: object
        "401":
          description: AUTH-003
          schema:
            allOf:
            - $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.Response'
            - properties:
                error:
                  $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.ErrorBody'
              type: object
        "403":
          description: AUTH-005
          schema:
            allOf:
            - $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.Response'
            - properties:
                error:
                  $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.ErrorBody'
              type: object
        "500":
          description: COMMON-901
          schema:
            allOf:
            - $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.Response'
            - properties:
                error:
                  $ref: '#/definitions/github_com_qwerty7415963_go_be_arbitrage_internal_api.ErrorBody'
              type: object
      security:
      - BearerAuth: []
      summary: Stop strategy instance
      tags:
      - strategies
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"